+++ src/UsdnProtocol/UsdnProtocolStorage.sol
@@ -156,8 +156,9 @@ abstract contract UsdnProtocolStorage is IUsdnProtocolStorage, InitializableReen
     /// @notice The fee collector's address
     address internal _feeCollector;
 
+    // TO DO
     /// @notice The deposit required for a new position (0.5 ether)
-    uint256 internal _securityDepositValue = 0.5 ether;
+    uint64 internal _securityDepositValue = 0.5 ether;
 
     /// @notice The nominal (target) price of USDN (with _priceFeedDecimals)
     uint128 internal _targetUsdnPrice;
@@ -412,7 +413,7 @@ abstract contract UsdnProtocolStorage is IUsdnProtocolStorage, InitializableReen
     }
 
     /// @inheritdoc IUsdnProtocolStorage
-    function getSecurityDepositValue() external view returns (uint256) {
+    function getSecurityDepositValue() external view returns (uint64) {
         return _securityDepositValue;
     }
